# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kartly REST API
  version: '1.0'
  description: |
    JSON API behind the Kartly storefront. The server-rendered pages and this API run the
    same service layer, so anything the website can do is available here.

    **Authentication** is a session cookie (`kartly.sid`), set by `POST /auth/login` or
    `POST /auth/register` and sent automatically by browsers.

    **CSRF:** every unsafe request (POST, PATCH, DELETE) must carry the session's token in an
    `X-CSRF-Token` header. Fetch one from `GET /auth/csrf-token`. The token is regenerated
    when the session is (that is, on sign-in and sign-out), so read it again from the login
    response.

    **Guest carts:** cart endpoints work without signing in. The cart is tied to the session
    and is merged into the user's cart on the next sign-in or registration.

    **Errors** always use the shape `{"error": {"code", "message", "details"?}}`.

    **Rate limits:** 900 requests per 15 minutes per IP for the API as a whole, and 20 per
    15 minutes for `/auth/login` and `/auth/register`.
  license:
    name: MIT
    url: https://github.com/VivekReddy2001/shopping_cart/blob/main/LICENSE
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://{host}/api/v1
    description: Deployed instance
    variables:
      host:
        default: kartly.onrender.com
...
paths:
...
  /products/{idOrSlug}:
    get:
      tags: [Catalogue]
      summary: One product by id or slug
      security: []
      parameters:
        - {
            name: idOrSlug,
            in: path,
            required: true,
            schema: { type: string },
            example: apple-iphone-11,
          }
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ProductDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /products/{id}:
    patch:
      tags: [Admin]
      summary: Update a product
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Any subset of the create fields. Omitted fields are left untouched.
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ProductDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Admin]
      summary: Archive a product
      description: |
        Products are never hard-deleted — `isActive` is set to `false` so past orders,
        reviews and analytics stay valid. The product disappears from the storefront.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Archived }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /orders/{orderNumber}:
    get:
      tags: [Orders]
      summary: One of your orders
      description: Orders belonging to someone else return 404, not 403, so the endpoint cannot be probed.
      parameters:
        - {
            name: orderNumber,
            in: path,
            required: true,
            schema: { type: string },
            example: KRT-260911-7Q4ZK,
          }
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Order' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /admin/orders:
    get:
      tags: [Admin]
      summary: Every order, with status counts
      parameters:
        - {
            name: status,
            in: query,
            schema: { type: string, enum: [PLACED, PROCESSING, SHIPPED, DELIVERED, CANCELLED] },
          }
        - {
            name: q,
            in: query,
            schema: { type: string },
            description: Order number,
            customer name or email,
          }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
      responses:
        '200': { description: Orders }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: kartly.sid
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        url: { type: string }
        image: { type: string }
        brand: { $ref: '#/components/schemas/Ref' }
        category: { $ref: '#/components/schemas/Ref' }
        price: { type: number }
        mrp: { type: number }
        discountPercent: { type: integer }
        stock: { type: integer }
        inStock: { type: boolean }
        rating:
          type: object
          properties:
            average: { type: number }
            count: { type: integer }
        featured: { type: boolean }
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            description: { type: string }
            highlights: { type: array, items: { type: string } }
            tags: { type: array, items: { type: string } }
            isActive: { type: boolean }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    Ref:
      type: object
      nullable: true
      properties:
        name: { type: string }
        slug: { type: string }
...
    Order:
      type: object
      properties:
        orderNumber: { type: string, example: KRT-260911-7Q4ZK }
        invoiceNumber: { type: string, example: INV-2026-000042 }
        status: { type: string, enum: [PLACED, PROCESSING, SHIPPED, DELIVERED, CANCELLED] }
        statusHistory:
          type: array
          items:
            type: object
            properties:
              status: { type: string }
              at: { type: string, format: date-time }
              note: { type: string }
        items:
          type: array
          items:
            type: object
            description: A snapshot taken at purchase time, not a live product reference.
            properties:
              productId: { type: string }
              name: { type: string }
              image: { type: string }
              price: { type: number }
              quantity: { type: integer }
              lineTotal: { type: number }
        shippingAddress: { $ref: '#/components/schemas/Address' }
        payment:
          type: object
          properties:
            method: { type: string, enum: [COD, ONLINE] }
            status: { type: string, enum: [PENDING, PAID, REFUNDED] }
            transactionId: { type: string, nullable: true }
            paidAt: { type: string, format: date-time, nullable: true }
        amounts:
          type: object
          properties:
            subtotal: { type: number }
            shipping: { type: number }
            tax: { type: number }
            total: { type: number }
        createdAt: { type: string, format: date-time }
        links:
          type: object
          properties:
            self: { type: string }
            invoicePdf: { type: string }
    Address:
      type: object
      required: [fullName, phone, line1, city, state, postalCode]
      properties:
        fullName: { type: string }
        phone: { type: string, example: '+91 98765 43210' }
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string, example: '560001' }
        country: { type: string, default: India }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: OUT_OF_STOCK }
            message: { type: string }
            details:
              type: object
              additionalProperties: { type: string }
              description: Field-level messages, for validation failures.
  responses:
...
    Forbidden:
      description: Signed in, but not allowed (including the read-only demo admin)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - sessionCookie: []
    csrfToken: []